refactor(sources/bigquery): move source implementation in Invoke() function to Source
